--- _examples/30graphql/go.mod
@@ -1,5 +1,0 @@
-module m
-
-go 1.17
-
-require github.com/graphql-go/graphql v0.8.0 // indirect
