+++ _examples/go.mod
@@ -12,6 +12,7 @@ require (
 	github.com/go-playground/universal-translator v0.18.0
 	github.com/go-playground/validator/v10 v10.9.0
 	github.com/gorilla/schema v1.2.0
+	github.com/graphql-go/graphql v0.8.0 // indirect
 	github.com/morikuni/failure v0.14.0
 	github.com/podhmo/apikit v0.0.0-20210923132720-b5e305e3689c
 	github.com/podhmo/reflect-openapi v0.0.13
